% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_utils.R
\name{combine_regions_from_files}
\alias{combine_regions_from_files}
\title{Combine Regions from Single TACs/Morph File Pair}
\usage{
combine_regions_from_files(
  derivatives_folder,
  tacs_relative_path,
  morph_relative_path,
  regions_for_files
)
}
\arguments{
\item{derivatives_folder}{Base path to derivatives folder}

\item{tacs_relative_path}{Relative path to TACs file from derivatives folder}

\item{morph_relative_path}{Relative path to morph file from derivatives folder}

\item{regions_for_files}{Filtered regions config for these specific files}
}
\value{
Tibble with all combined regions for this file pair
}
\description{
Process all region combinations for one TACs/morph file pair
}
