% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_functions.R
\name{execute_pipeline_step}
\alias{execute_pipeline_step}
\title{Execute Individual Pipeline Step}
\usage{
execute_pipeline_step(step, analysis_folder, bids_dir, blood_dir, config)
}
\arguments{
\item{step}{Character string step name}

\item{analysis_folder}{Character string path to analysis folder}

\item{bids_dir}{Character string path to BIDS directory (can be NULL)}

\item{blood_dir}{Character string path to blood directory (can be NULL)}

\item{config}{List containing petfit configuration}
}
\value{
List with step execution result
}
\description{
Execute a single step of the petfit pipeline
}
