% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_utils.R
\name{extract_pet_id_from_kinpar_filename}
\alias{extract_pet_id_from_kinpar_filename}
\title{Extract PET ID from Kinpar Filename}
\usage{
extract_pet_id_from_kinpar_filename(filename)
}
\arguments{
\item{filename}{Basename of a kinpar file}
}
\value{
Character string PET ID
}
\description{
Extract the PET measurement identifier from a kinpar filename.
The PET ID is the portion before any \emph{model} or _desc- prefix.
}
