% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_utils.R
\name{find_tacs_folders}
\alias{find_tacs_folders}
\title{Find Folders Containing TACs Files}
\usage{
find_tacs_folders(derivatives_folder)
}
\arguments{
\item{derivatives_folder}{Character string path to the derivatives folder}
}
\value{
Character vector of directory paths containing *_tacs.tsv files
}
\description{
Identifies directories that contain *_tacs.tsv files
}
