% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_fitting.R
\name{fit_single_measurement_ref}
\alias{fit_single_measurement_ref}
\title{Fit a single reference-tissue measurement interactively}
\usage{
fit_single_measurement_ref(
  analysis_folder,
  model_number,
  pet,
  region,
  ancillary_path = NULL
)
}
\arguments{
\item{analysis_folder}{Path to the analysis folder.}

\item{model_number}{Which configured model to fit (see plasma version).}

\item{pet}{PET measurement identifier.}

\item{region}{Region to fit.}

\item{ancillary_path}{Optional ancillary analysis folder (for k2prime from
an ancillary analysis).}
}
\value{
A list with \code{fit}, \code{par}, \code{par.se}, \code{type}, \code{model_number}, \code{pet}, \code{region}.
}
\description{
Reference-tissue counterpart of \code{\link[=fit_single_measurement_plasma]{fit_single_measurement_plasma()}},
replicating \code{mrtm1_report.Rmd}, \code{mrtm2_report.Rmd}, \code{reflogan_report.Rmd},
\code{srtm_report.Rmd} and \code{srtm2_report.Rmd}.
}
