% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_generation.R
\name{get_model_template}
\alias{get_model_template}
\title{Get Model Template Name}
\usage{
get_model_template(model_type)
}
\arguments{
\item{model_type}{Character string type of model}
}
\value{
Character string template filename
}
\description{
Map model type to corresponding template filename
}
