% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary_utils.R
\name{scan_ancillary_contents}
\alias{scan_ancillary_contents}
\title{Scan Ancillary Folder Contents}
\usage{
scan_ancillary_contents(ancillary_path)
}
\arguments{
\item{ancillary_path}{Full path to the ancillary analysis folder}
}
\value{
List with available contents: delay_files, model1_kinpar, model2_kinpar, model3_kinpar
}
\description{
Scan an ancillary analysis folder to determine what outputs are available
(delay files, model kinpar files).
}
