% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_utils.R
\name{attributes_to_title}
\alias{attributes_to_title}
\title{Extract a pet variable from BIDS attributes}
\usage{
attributes_to_title(bidsdata, all_attributes = FALSE)
}
\arguments{
\item{bidsdata}{The result of bids_parse_files()}

\item{all_attributes}{Make a column including attributes which are all the same.}
}
\value{
A vector of pet measurement identifiers specific to each PET.
}
\description{
Extract a pet variable from BIDS attributes
}
\examples{
\dontrun{
studydata <- bids_parse_files(studypath)
studydata$pet <- attributes_to_title(studydata)
}
}
