% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_utils.R
\name{create_combined_tacs_json_description}
\alias{create_combined_tacs_json_description}
\title{Create JSON Description File for Combined TACs}
\usage{
create_combined_tacs_json_description(
  participants_metadata,
  injected_radioactivity_units,
  original_tac_units = "Bq",
  output_dir
)
}
\arguments{
\item{participants_metadata}{Participants metadata from participants.json (can be NULL)}

\item{injected_radioactivity_units}{Units for InjectedRadioactivity (can be NA)}

\item{original_tac_units}{Original radioactivity units detected from source TACs files (default "Bq")}

\item{output_dir}{Directory to save the JSON file}
}
\value{
Path to created JSON file
}
\description{
Generate JSON metadata file describing columns in combined TACs file
}
