% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petfit-package.R
\docType{package}
\name{petfit-package}
\alias{petfit}
\alias{petfit-package}
\title{petfit: PETFit BIDS App Configuration Interface}
\description{
PETFit is a BIDS app and Shiny UI for performing PET kinetic modelling using kinfitr. It provides a user interface for configuring kinetic modelling parameters for time–activity curves (TACs).
}
\author{
\strong{Maintainer}: Granville Matheson \email{granville.matheson@ki.se}

Other contributors:
\itemize{
  \item Alessio Giacomel \email{alessio.giacomel@kcl.ac.uk} [contributor]
  \item Martin Noergaard \email{martin.noergaard@di.ku.dk} [contributor]
  \item Anthony Galassi \email{anthony.galassi@nih.gov} [contributor]
}

}
\keyword{internal}
