% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting_utils.R
\name{subset_combined_tacs}
\alias{subset_combined_tacs}
\title{Subset Combined TACs Data}
\usage{
subset_combined_tacs(combined_tacs_data, subset_params)
}
\arguments{
\item{combined_tacs_data}{Tibble with combined TACs data}

\item{subset_params}{List of subsetting parameters}
}
\value{
Filtered tibble
}
\description{
Filter combined TACs data based on subsetting criteria
}
